Renamed sendTroopsShips to stationArmy
